fix: split waste contract definitions
