+++ packages/waste-management-contracts/src/index.ts
@@ -1,1 +1,2 @@
+export * from './import-profile-definitions.js';
 export * from './job-definitions.js';
